Bluetooth: Host: add bt_id_delete_all function to remove all ids
All ids including BT_ID_DEFAULT can be removed with this function.

This is called from bt_disable as BT_ID_DEFAULT is initialized in
bt_enable. We assert this is successful as so much of the disable has
happened already, returning the error would be of no use.

Fixed a bug when you call bt_unpair with NULL it would try and access
this memory.

Signed-off-by: Sean Madigan <sean.madigan@nordicsemi.no>